ApplyPileupsParam {Rsamtools}	R Documentation

Parameters for creating pileups from BAM files

Description

Use ApplyPileupsParam() to create a parameter object influencing what fields and which records are used to calculate pile-ups, and to influence the values returned.

Usage

# Constructor
ApplyPileupsParam(flag = scanBamFlag(),
    minBaseQuality = 13L, minMapQuality = 0L,
    minDepth = 0L, maxDepth = 250L,
    yieldSize = 1L, yieldBy = c("range", "position"), yieldAll = FALSE,
    which = GRanges(), what = c("seq", "qual"))

# Accessors
plpFlag(object)
plpFlag(object) <- value
plpMaxDepth(object)
plpMaxDepth(object) <- value
plpMinBaseQuality(object)
plpMinBaseQuality(object) <- value
plpMinDepth(object)
plpMinDepth(object) <- value
plpMinMapQuality(object)
plpMinMapQuality(object) <- value
plpWhat(object)
plpWhat(object) <- value
plpWhich(object)
plpWhich(object) <- value
plpYieldAll(object)
plpYieldAll(object) <- value
plpYieldBy(object)
plpYieldBy(object) <- value
plpYieldSize(object)
plpYieldSize(object) <- value

## S4 method for signature 'ApplyPileupsParam'
show(object)

Arguments

flag
An instance of the object returned by scanBamFlag, restricting various aspects of reads to be included or excluded.

minBaseQuality
The minimum read base quality below which the base is ignored when summarizing pileup information.

minMapQuality
The minimum mapping quality below which the entire read is ignored.

minDepth
The minimum depth of the pile-up below which the position is ignored.

maxDepth
The maximum depth of reads considered at any position; this can be used to limit memory consumption.

yieldSize
The number of records to include in each call to FUN.

yieldBy
How records are to be counted. By range (in which case yieldSize must equal 1) means that FUN is invoked once for each range in which. By position means that FUN is invoked whenever pile-ups have been accumulated for yieldSize positions, regardless of ranges in which.

yieldAll
Whether to report all positions (yieldAll=TRUE), or just those passing the filtering criteria of flag, minBaseQuality, etc. When yieldAll=TRUE, positions not passing filter criteria have ‘0’ entries in seq or qual.

which
A GRanges or IntegerRangesList instance restricting pileup calculations to the corresponding genomic locations.

what
A character() instance indicating what values are to be returned. One or more of c("seq", "qual").

object
An instace of class ApplyPileupsParam.

value
An instance to be assigned to the corresponding slot of the ApplyPileupsParam instance.

Objects from the Class

Objects are created by calls of the form ApplyPileupsParam().

Slots

Slot interpretation is as described in the ‘Arguments’ section.

flag

Object of class integer encoding flags to be kept when they have their '0' (keep0) or '1' (keep1) bit set.

minBaseQuality

An integer(1).

minMapQuality

An integer(1).

minDepth

An integer(1).

maxDepth

An integer(1).

yieldSize

An integer(1).

yieldBy

An character(1).

yieldAll

A logical(1).

which

A GRanges or IntegerRangesList object.

what

A character().

Functions and methods

See 'Usage' for details on invocation.

Constructor:

ApplyPileupsParam:

Returns a ApplyPileupsParam object.

Accessors: get or set corresponding slot values; for setters, value is coerced to the type of the corresponding slot.

plpFlag, plpFlag<-

Returns or sets the named integer vector of flags; see scanBamFlag.

plpMinBaseQuality, plpMinBaseQuality<-

Returns or sets an integer(1) vector of miminum base qualities.

plpMinMapQuality, plpMinMapQuality<-

Returns or sets an integer(1) vector of miminum map qualities.

plpMinDepth, plpMinDepth<-

Returns or sets an integer(1) vector of miminum pileup depth.

plpMaxDepth, plpMaxDepth<-

Returns or sets an integer(1) vector of the maximum depth to which pileups are calculated.

plpYieldSize, plpYieldSize<-

Returns or sets an integer(1) vector of yield size.

plpYieldBy, plpYieldBy<-

Returns or sets an character(1) vector determining how pileups will be returned.

plpYieldAll, plpYieldAll<-

Returns or sets an logical(1) vector indicating whether all positions, or just those satisfying pileup positions, are to be returned.

plpWhich, plpWhich<-

Returns or sets the object influencing which locations pileups are calculated over.

plpWhat, plpWhat<-

Returns or sets the character vector describing what summaries are returned by pileup.

Methods:

show

Compactly display the object.

Author(s)

Martin Morgan

See Also

applyPileups.

Examples

example(applyPileups)

[Package Rsamtools version 2.6.0 Index]