RsamtoolsFile {Rsamtools}	R Documentation

A base class for managing file references in Rsamtools

Description

RsamtoolsFile is a base class for managing file references in Rsamtools; it is not intended for direct use by users – see, e.g., BamFile.

Usage

## accessors
## S4 method for signature 'RsamtoolsFile'
path(object, ...)
## S4 method for signature 'RsamtoolsFile'
index(object, ..., asNA = TRUE)
## S4 method for signature 'RsamtoolsFile'
isOpen(con, rw="")
## S4 method for signature 'RsamtoolsFile'
yieldSize(object, ...)
yieldSize(object, ...) <- value
## S4 method for signature 'RsamtoolsFile'
show(object)

Arguments

con, object
An instance of a class derived from RsamtoolsFile.

asNA
logical indicating if missing output should be NA or character()

rw
Mode of file; ignored.

...
Additional arguments, unused.

value
Replacement value.

Objects from the Class

Users do not directly create instances of this class; see, e.g., BamFile-class.

Fields

The RsamtoolsFile class is implemented as an S4 reference class. It has the following fields:

.extptr

An externalptr initialized to an internal structure with opened bam file and bam index pointers.

path

A character(1) vector of the file name.

index

A character(1) vector of the index file name.

yieldSize

An integer(1) vector of the number of records to yield.

Functions and methods

Accessors:

path

Returns a character(1) vector of path names.

index

Returns a character(1) vector of index path names.

yieldSize, yieldSize<-

Return or set an integer(1) vector indicating yield size.

Methods:

isOpen

Report whether the file is currently open.

show

Compactly display the object.

Author(s)

Martin Morgan

[Package Rsamtools version 2.6.0 Index]