RsamtoolsFileList {Rsamtools}	R Documentation

A base class for managing lists of Rsamtools file references

Description

RsamtoolsFileList is a base class for managing lists of file references in Rsamtools; it is not intended for direct use – see, e.g., BamFileList.

Usage

## S4 method for signature 'RsamtoolsFileList'
path(object, ...)
## S4 method for signature 'RsamtoolsFileList'
index(object, ..., asNA = TRUE)
## S4 method for signature 'RsamtoolsFileList'
isOpen(con, rw="")
## S3 method for class 'RsamtoolsFileList'
open(con, ...)
## S3 method for class 'RsamtoolsFileList'
close(con, ...)
## S4 method for signature 'RsamtoolsFileList'
names(x)
## S4 method for signature 'RsamtoolsFileList'
yieldSize(object, ...)

Arguments

con, object, x
An instance of a class derived from RsamtoolsFileList.

asNA
logical indicating if missing output should be NA or character()

rw
Mode of file; ignored.

...
Additional arguments.

Objects from the Class

Users do not directly create instances of this class; see, e.g., BamFileList-class.

Functions and methods

This class inherits functions and methods for subseting, updating, and display from the SimpleList class.

Methods:

isOpen:

Report whether each file in the list is currently open.

open:

Attempt to open each file in the list.

close:

Attempt to close each file in the list.

names:

Names of each element of the list or, if names are NULL, the basename of the path of each element.

Author(s)

Martin Morgan

[Package Rsamtools version 2.6.0 Index]