ScanBcfParam-class {Rsamtools}	R Documentation

Parameters for scanning BCF files

Description

Use ScanBcfParam() to create a parameter object influencing the ‘INFO’ and ‘GENO’ fields parsed, and which sample records are imported from a BCF file. Use of which requires that a BCF index file (<filename>.bci) exists.

Usage

ScanBcfParam(fixed=character(), info=character(), geno=character(),
             samples=character(), trimEmpty=TRUE, which, ...)

## S4 method for signature 'missing'
ScanBcfParam(fixed=character(), info=character(), geno=character(), 
             samples=character(), trimEmpty=TRUE, which, ...)
## S4 method for signature 'IntegerRangesList'
ScanBcfParam(fixed=character(), info=character(), geno=character(), 
             samples=character(), trimEmpty=TRUE, which, ...)
## S4 method for signature 'GRanges'
ScanBcfParam(fixed=character(), info=character(), geno=character(), 
             samples=character(), trimEmpty=TRUE, which, ...)
## S4 method for signature 'GRangesList'
ScanBcfParam(fixed=character(), info=character(), geno=character(), 
             samples=character(), trimEmpty=TRUE, which, ...)

## Accessors
bcfFixed(object)
bcfInfo(object)
bcfGeno(object)
bcfSamples(object)
bcfTrimEmpty(object)
bcfWhich(object)

Arguments

fixed
A logical(1) for use with ScanVcfParam only.

info
A character() vector of ‘INFO’ fields (see scanVcfHeader) to be returned.

geno
A character() vector of ‘GENO’ fields (see scanVcfHeader) to be returned. character(0) returns all fields, NA_character_ returns none.

samples
A character() vector of sample names (see scanVcfHeader) to be returned. character(0) returns all fields, NA_character_ returns none.

trimEmpty
A logical(1) indicating whether ‘GENO’ fields with no values should be returned.

which
An object, for which a method is defined (see usage, above), describing the sequences and ranges to be queried. Variants whose POS lies in the interval(s) [start, end) are returned.

object
An instance of class ScanBcfParam.

...
Arguments used internally.

Objects from the Class

Objects can be created by calls of the form ScanBcfParam().

Slots

which:

Object of class "IntegerRangesList" indicating which reference sequence and coordinate variants must overlap.

info:

Object of class "character" indicating portions of ‘INFO’ to be returned.

geno:

Object of class "character" indicating portions of ‘GENO’ to be returned.

samples:

Object of class "character" indicating the samples to be returned.

trimEmpty:

Object of class "logical" indicating whether empty ‘GENO’ fields are to be returned.

fixed:

Object of class "character". For use with ScanVcfParam only.

Functions and methods

See 'Usage' for details on invocation.

Constructor:

ScanBcfParam:

Returns a ScanBcfParam object. The which argument to the constructor can be one of several types, as documented above.

Accessors:

bcfInfo, bcfGeno, bcfTrimEmpty, bcfWhich:

Return the corresponding field from object.

Methods:

show

Compactly display the object.

Author(s)

Martin Morgan mtmorgan@fhcrc.org

See Also

scanVcf ScanVcfParam

Examples

## see ?ScanVcfParam examples

[Package Rsamtools version 2.6.0 Index]