headerTabix {Rsamtools}	R Documentation

Retrieve sequence names defined in a tabix file.

Description

This function queries a tabix file, returning the names of the ‘sequences’ used as a key when creating the file.

Usage

headerTabix(file, ...)
## S4 method for signature 'character'
headerTabix(file, ...)

Arguments

file
A character(1) file path or TabixFile instance pointing to a ‘tabix’ file.

...
Additional arguments, currently ignored.

Value

A list(4) of the sequence names, column indicies used to sort the file, the number of lines skipped while indexing, and the comment character used while indexing.

Author(s)

Martin Morgan <mtmorgan@fhcrc.org>.

Examples

fl <- system.file("extdata", "example.gtf.gz", package="Rsamtools",
                  mustWork=TRUE)
headerTabix(fl)

[Package Rsamtools version 2.6.0 Index]