TabixInput {Rsamtools}	R Documentation

Operations on ‘tabix’ (indexed, tab-delimited) files.

Description

Scan compressed, sorted, tabix-indexed, tab-delimited files.

Usage

scanTabix(file, ..., param)
## S4 method for signature 'character,IntegerRangesList'
scanTabix(file, ..., param)
## S4 method for signature 'character,GRanges'
scanTabix(file, ..., param)

Arguments

file
The character() file name(s) of the tabix file be processed, or more flexibly an instance of class TabixFile.

param
A instance of GRanges or IntegerRangesList providing the sequence names and regions to be parsed.

...
Additional arguments, currently ignored.

Value

scanTabix returns a list, with one element per region. Each element of the list is a character vector representing records in the region.

Error

scanTabix signals errors using signalCondition. The following errors are signaled:

scanTabix_param

yieldSize(file) must be NA when more than one range is specified.

scanTabix_io

A read error occured while inputing the tabix file. This might be because the file is corrupt, or of incorrect format (e.g., when path points to a plain text file but index is present, implying that path should be a bgziped file. The error message may include an error code representing the logical OR of these cryptic signals: 1, BGZF_ERR_ZLIB; 2, BGZF_ERR_HEADER; 4, BGZF_ERR_IO; 8, BGZF_ERR_MISUSE.

Author(s)

Martin Morgan <mtmorgan@fhcrc.org>.

References

http://samtools.sourceforge.net/tabix.shtml

Examples

example(TabixFile)

[Package Rsamtools version 2.6.0 Index]