testPairedEndBam {Rsamtools}	R Documentation

Quickly test if a BAM file has paired end reads

Description

Iterate through a BAM file until a paired-end read is encountered or the end of file is reached; report the occurrence of paired-end reads to the user.

Usage

testPairedEndBam(file, index=file, ...)

Arguments

file
character(1) BAM file name, or a BamFile instance. Open BamFiles are closed; their yield size is respected when iterating through the file.

index
(optional) character(1) name of the index file of the 'BAM' file being processed; this is given without the '.bai' extension.

...
Additional arguments, currently unused.

Value

A logical vector of length 1 containing TRUE is returned if BAM file contained paired end reads, FALSE otherwise.

Author(s)

Martin Morgan mailto:mtmorgan@fhcrc.org, Sonali Arora mailto:sarora@fhcrc.org

Examples

fl <- system.file("extdata", "ex1.bam", package="Rsamtools")
testPairedEndBam(fl)

[Package Rsamtools version 2.6.0 Index]